requirements.txt
streamlit
google-generativeai
matplotlib
numpy
python-dotenv
